Address frontend comments
