+++ packages/website/src/store/Reefs/helpers.ts
@@ -2,8 +2,6 @@ import { isNil } from "lodash";
 
 import type { TableRow } from "../Homepage/types";
 import type {
-  CollectionData,
-  CollectionDataResponse,
   Metrics,
   MetricsKeys,
   OceanSenseData,
@@ -61,23 +59,6 @@ export const constructTableData = (list: Reef[]): TableRow[] => {
     };
   });
 };
-
-export const mapCollectionData = (
-  data: CollectionDataResponse
-): CollectionData => ({
-  alert: data.alert,
-  bottomTemperature: data.bottom_temperature,
-  dhw: data.dhw,
-  satelliteTemperature: data.satellite_temperature,
-  sstAnomaly: data.sst_anomaly,
-  topTemperature: data.top_temperature,
-  significantWaveHeight: data.significant_wave_height,
-  waveMeanDirection: data.wave_mean_direction,
-  wavePeakPeriod: data.wave_peak_period,
-  windDirection: data.wind_direction,
-  windSpeed: data.wind_speed,
-  weeklyAlert: data.weekly_alert,
-});
 
 const mapMetrics = <T>(
   data: Record<MetricsKeys, T[]>
